Revert "Code refactor- split ambiguous routes.js"
