+++ routes/index.js
@@ -1,6 +1,6 @@
 var express = require("express");
 var router = express.Router();
-const { homepage } = require("../controllers/meme")
+const { homepage } = require("../controllers/routes");
 /* GET home page. */
 router.get("/", homepage);
