+++ routes/users.js
@@ -2,7 +2,7 @@ var express = require("express");
 var router = express.Router();
 const jwt = require("jsonwebtoken");
 const Users = require("../model/users");
-const { getUsers, login, register } = require("../controllers/user");
+const { getUsers, login, register } = require("../controllers/routes");
 /* GET users listing. */
 router.get("/", getUsers);
